Microservices swarm

+++ src/auth_service/auth_service/auth_service/db/init_db.py
@@ -8,8 +8,9 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from auth_service import crud, schemas
 from auth_service.core.config import get_app_settings
-from auth_service.db.session import Base, MasterSlaves, scoped_session
-from auth_service.db.session import engines
+from auth_service.db.base import ReplType
+from auth_service.db.session import Base, MasterReplica, scoped_session
+from auth_service.db.session import async_engines
 from auth_service.models import UserRole
 from auth_service.models.user_role import UserRoleEnum
 
@@ -29,37 +30,38 @@ async def execute_sql_file(
 
 
 async def base_metadata(
-    ms: MasterSlaves, create: bool = False, drop: bool = False
+    ms: MasterReplica, create: bool = False, drop: bool = False
 ):
     for _type, _eng in ms.engine.items():
-        for eng in _eng:
-            msg = f"engine_type: {_type.name}, url: {eng.url}"
-            try:
-                async with eng.begin() as conn:
-                    if drop:
-                        action = "metadata.drop_all"
-                        await conn.run_sync(Base.metadata.drop_all)
-                    elif create:
-                        action = "metadata.create_all"
-                        await conn.run_sync(Base.metadata.create_all)
-                    else:
-                        raise ValueError(
-                            "'drop' or 'create' must be True", drop, create
+        if _type == ReplType.master:
+            for eng in _eng:
+                msg = f"repl_type: {_type.name}, url: {eng.url}"
+                action = ""
+                try:
+                    async with eng.begin() as conn:
+                        if drop:
+                            action = "metadata.drop_all"
+                            await conn.run_sync(Base.metadata.drop_all)
+                        elif create:
+                            action = "metadata.create_all"
+                            await conn.run_sync(Base.metadata.create_all)
+                        else:
+                            raise ValueError(
+                                "'drop' or 'create' must be True", drop, create
+                            )
+                        logger.opt(colors=True).info(
+                            f"<fg 255,70,230>{action}</fg 255,70,230> - {msg}"
                         )
-                logger.opt(colors=True).info(
-                    f"<fg 255,70,230>{action}</fg 255,70,230> - {msg}"
-                )
-            except ConnectionRefusedError as ex:
-                logger.opt(colors=True).error(
-                    f"<fg 255,70,230>{action}</fg 255,70,230> - {msg}, {ex.__class__.__name__} {ex}"
-                )
+                except ConnectionRefusedError as ex:
+                    logger.opt(colors=True).error(
+                        f"<fg 255,70,230>{action}</fg 255,70,230> - {msg}, {ex.__class__.__name__} {ex}"
+                    )
 
 
 async def create_roles(db: AsyncSession):
     for r in UserRoleEnum:
         if not await db.get(UserRole, r.name):
             await db.execute(insert(UserRole).values(name=r.name))
-    await db.commit()
 
 
 async def create_first_superuser(db: AsyncSession):
@@ -81,15 +83,16 @@ async def create_first_superuser(db: AsyncSession):
         )
         super_user = await crud.user.create(db, obj_in=user_in_admin)
         logger.info("created")
-    logger.info("already exists")
+    else:
+        logger.info("already exists")
     return super_user
 
 
 async def init_db():
     try:
-        await engines.check_engines()
-        await base_metadata(engines, create=True)
+        await async_engines.check_engines()
+        await base_metadata(async_engines, create=True)
         async with scoped_session() as db:
             await create_first_superuser(db)
     except Exception as e:
-        logger.error(f"{e.__class__.__name__} {e}")
+        logger.exception(e)
